feat(datasets): add initial interfaces for datasets authz
