+++ workbench-core/datasets/src/models/addRemoveAccessPermissionRequest.ts
@@ -1,0 +1,8 @@
+import { DataSetPermission } from './dataSetPermission';
+
+export interface AddRemoveAccessPermissionRequest {
+  /** the ID of the dataset */
+  dataSetId: string;
+  /** the permission to add or remove */
+  permission: DataSetPermission;
+}
