+++ workbench-core/datasets/src/models/dataSetsAccessLevel.ts
@@ -1,0 +1,1 @@
+export type DataSetsAccessLevel = 'read-only' | 'read-write';
