+++ workbench-core/datasets/src/models/getAccessPermissionRequest.ts
@@ -1,0 +1,6 @@
+export interface GetAccessPermissionRequest {
+  /** the ID of the dataset */
+  dataSetId: string;
+  /** the user or group for which permissions are sought */
+  subject: string;
+}
